feature: description defaults
Added the ability to use default reasons for why a skill is demonstrated, by putting null. A skill can have its own default. If a skill isn't given a default, the categories' default is used instead. If a category(and none of its parents) have defaults, an error is thrown.